feat(sidebar): update layout to use 3 columns
besides, this commit will also manage to clamp the sidebar text to workaround long category names.
In order for that to happen is needed to HACK the width with a fixed number that will avoid further
complexity. Far from ideal but avoid more complexity that may not be worth it now.
https://www.loom.com/share/baf92bd0590d464eba0dd1d1173a7cc0